+++ src/ArduinoJson/Array/ArrayRef.hpp
@@ -96,10 +96,6 @@ class ArrayConstRef : public ArrayRefBase<const CollectionData>,
   }
 
   FORCE_INLINE VariantConstRef operator[](size_t index) const {
-    return getElementConst(index);
-  }
-
-  FORCE_INLINE VariantConstRef getElementConst(size_t index) const {
     return VariantConstRef(_data ? _data->getElement(index) : 0);
   }
 };
@@ -152,21 +148,6 @@ class ArrayRef : public ArrayRefBase<CollectionData>,
     return ArrayConstRef(_data) == ArrayConstRef(rhs._data);
   }
 
-  // Internal use
-  FORCE_INLINE VariantRef getOrAddElement(size_t index) const {
-    return VariantRef(_pool, _data ? _data->getOrAddElement(index, _pool) : 0);
-  }
-
-  // Gets the value at the specified index.
-  FORCE_INLINE VariantRef getElement(size_t index) const {
-    return VariantRef(_pool, _data ? _data->getElement(index) : 0);
-  }
-
-  // Gets the value at the specified index.
-  FORCE_INLINE VariantConstRef getElementConst(size_t index) const {
-    return VariantConstRef(_data ? _data->getElement(index) : 0);
-  }
-
   // Removes element at specified position.
   FORCE_INLINE void remove(iterator it) const {
     if (!_data)
@@ -185,6 +166,18 @@ class ArrayRef : public ArrayRefBase<CollectionData>,
     if (!_data)
       return;
     _data->clear();
+  }
+
+  MemoryPool* getPool() const {
+    return _pool;
+  }
+
+  VariantData* getData() const {
+    return collectionToVariant(_data);
+  }
+
+  VariantData* getOrCreateData() const {
+    return collectionToVariant(_data);
   }
 
  private: