+++ src/ArduinoJson/Object/ObjectFunctions.hpp
@@ -22,14 +22,4 @@ void objectRemove(CollectionData *obj, TAdaptedString key) {
     return;
   obj->removeMember(key);
 }
-
-template <typename TAdaptedString, typename TStoragePolicy>
-inline VariantData *objectGetOrAddMember(CollectionData *obj,
-                                         TAdaptedString key, MemoryPool *pool,
-                                         TStoragePolicy storage_policy) {
-  if (!obj)
-    return 0;
-
-  return obj->getOrAddMember(key, pool, storage_policy);
-}
 }  // namespace ARDUINOJSON_NAMESPACE